requirements.txt
alive_progress==2.4.1
colorama==0.4.5
dnspython==2.2.1
prettytable==2.1.0
